test: split test-cli-syntax into multiple tests
Split test-cli-syntax into multiple files to improve reliability and/or
isolate unreliable test cases.

Move test cases back to parallel as appropriate.
